requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deepface==0.0.93
defusedxml==0.7.1
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
gast==0.4.0
gdown==5.2.0
Glances==4.1.2
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==2.12.0
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
MarkupSafe==3.0.1
ml_dtypes==0.5.0
mtcnn==1.0.0
# numpy==1.23.5
nvidia-ml-py==12.560.30
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
psutil==6.1.0
py3nvml==0.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
scipy==1.14.1
six==1.16.0
soupsieve==2.6
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.14.1
xmltodict==0.14.2
